+++ lib/core/extensions/t_html_dom_extensions.dart
@@ -1,0 +1,9 @@
+import 'package:html/dom.dart' as html;
+
+extension THtmlDomExtension on html.Document {
+  void cleanDomTag({
+    String tagNames = 'script,style,noscript, div[id^="pf-"]',
+  }) {
+    querySelectorAll(tagNames).forEach((e) => e.remove());
+  }
+}