+++ lib/core/index.dart
@@ -1,5 +1,5 @@
 export 'f_query.dart';
 export 't_extractor.dart';
-export 't_html_extensions.dart';
+export 'extensions/t_html_dom_extensions.dart';
 export 't_selector_rules.dart';
 export 't_html_parser_core.dart';